refactor main into smaller files
