+++ main.ts
@@ -1,14 +1,8 @@
-import { App, MarkdownView, Plugin, PluginManifest, Modal } from "obsidian";
+import { App, MarkdownView, Plugin, PluginManifest } from "obsidian";
 import { StatBarSettingTab, MyPluginSettings, DEFAULT_SETTINGS } from "./src/settings";
-import { debugLog } from "./src/debug";
-
-interface DocumentStats {
-	wordCount: number;
-	charCount: number;
-	readTime: string;
-	isSelection?: boolean;
-	selectionPrefix?: string;
-}
+import { DocumentStats } from "./src/types";
+import { getWordCount, calculateReadTime } from "./src/stats";
+import { DetailedStatsModal } from "./src/modal";
 
 export default class StatBarPlugin extends Plugin {
 	settings!: MyPluginSettings; // Use definite assignment assertion
@@ -127,8 +121,8 @@ export default class StatBarPlugin extends Plugin {
 				wordCount = cachedStats.wordCount;
 				readTime = cachedStats.readTime;
 			} else {
-				wordCount = this.getWordCount(text);
-				readTime = this.calculateReadTime(wordCount);
+				wordCount = getWordCount(text);
+				readTime = calculateReadTime(wordCount, this.settings.wordsPerMinute);
 
 				// Cache the results (separate caching for selection vs document)
 				const stats: DocumentStats = {
@@ -199,54 +193,6 @@ export default class StatBarPlugin extends Plugin {
 		}
 	}
 
-	public getWordCount(text: string): number {
-		debugLog("Raw input text:", text);
-
-		// Step-by-step cleaning for better accuracy
-		let cleanText = text;
-
-		// Remove code blocks first (multiline)
-		cleanText = cleanText.replace(/```[\s\S]*?```/g, "");
-
-		// Remove inline code
-		cleanText = cleanText.replace(/`[^`]*`/g, "");
-
-		// Process wiki links - extract content
-		cleanText = cleanText.replace(/\[\[([^\]]+)\]\]/g, "$1");
-
-		// Process markdown links - extract link text
-		cleanText = cleanText.replace(/\[([^\]]+)\]\([^)]+\)/g, "$1");
-
-		// Remove markdown syntax characters
-		cleanText = cleanText.replace(/[#*_~>]/g, "");
-
-		// Remove punctuation that might be attached to words
-		cleanText = cleanText.replace(/[.,!?;:]/g, " ");
-
-		// Normalize whitespace
-		cleanText = cleanText.replace(/\s+/g, " ").trim();
-
-		debugLog("After cleaning:", cleanText);
-
-		// Final word count calculation
-		if (!cleanText) return 0;
-		const words = cleanText.split(/\s+/).filter((word) => word.length > 0);
-		debugLog("Words array:", words);
-		debugLog("Word count:", words.length);
-
-		return words.length;
-	}
-
-	public calculateReadTime(wordCount: number): string {
-		const totalSeconds = Math.round(
-			(wordCount / this.settings.wordsPerMinute) * 60
-		);
-		const minutes = Math.floor(totalSeconds / 60);
-		const seconds = totalSeconds % 60;
-
-		return `${minutes}:${seconds.toString().padStart(2, "0")}`; // Format as "MM:SS"
-	}
-
 	public updateLastSavedTime() {
 		if (this.settings.showLastSavedTime) {
 			const now = new Date();
@@ -305,113 +251,5 @@ export default class StatBarPlugin extends Plugin {
 
 	private showDetailedStatsModal(): void {
 		new DetailedStatsModal(this.app, this).open();
-	}
-}
-
-class DetailedStatsModal extends Modal {
-	plugin: StatBarPlugin;
-
-	constructor(app: App, plugin: StatBarPlugin) {
-		super(app);
-		this.plugin = plugin;
-	}
-
-	onOpen() {
-		const { contentEl } = this;
-		contentEl.empty();
-
-		// Get current stats
-		const activeView = this.app.workspace.getActiveViewOfType(MarkdownView);
-		if (!activeView) {
-			contentEl.createEl('p', { text: 'No active markdown file' });
-			return;
-		}
-
-		const editor = activeView.editor;
-		const fromCursor = editor.getCursor('from');
-		const toCursor = editor.getCursor('to');
-		const hasSelection = fromCursor.line !== toCursor.line || fromCursor.ch !== toCursor.ch;
-		const selectedText = hasSelection ? editor.getSelection() : '';
-		const fullText = activeView.getViewData();
-		const isSelection = hasSelection && selectedText.length > 0;
-
-		// Calculate stats for both selection and full document
-		const currentText = isSelection ? selectedText : fullText;
-		const wordCount = this.plugin.getWordCount(currentText);
-		const charCount = currentText.length;
-		const charNoSpaces = currentText.replace(/\s/g, '').length;
-		const readTime = this.plugin.calculateReadTime(wordCount);
-
-		// Current scope indicator (serves as main header)
-		if (isSelection) {
-			const scopeEl = contentEl.createEl('div', { cls: 'statbar-modal-scope' });
-			scopeEl.createEl('strong', { text: '📝 Selection Statistics' });
-			scopeEl.createEl('p', { text: `Analyzing selected text (${selectedText.length} characters)` });
-		} else {
-			const scopeEl = contentEl.createEl('div', { cls: 'statbar-modal-scope' });
-			scopeEl.createEl('strong', { text: '📄 Document Statistics' });
-			scopeEl.createEl('p', { text: 'Analyzing entire document' });
-		}
-
-		// Stats container
-		const statsContainer = contentEl.createEl('div', { cls: 'statbar-modal-stats' });
-
-		// Word count
-		const wordEl = statsContainer.createEl('div', { cls: 'statbar-stat-item' });
-		wordEl.createEl('span', { text: 'Words: ', cls: 'statbar-stat-label' });
-		wordEl.createEl('span', { text: wordCount.toLocaleString(), cls: 'statbar-stat-value' });
-
-		// Character count
-		const charEl = statsContainer.createEl('div', { cls: 'statbar-stat-item' });
-		charEl.createEl('span', { text: 'Characters: ', cls: 'statbar-stat-label' });
-		charEl.createEl('span', { text: charCount.toLocaleString(), cls: 'statbar-stat-value' });
-
-		// Character count (no spaces)
-		const charNoSpacesEl = statsContainer.createEl('div', { cls: 'statbar-stat-item' });
-		charNoSpacesEl.createEl('span', { text: 'Characters (no spaces): ', cls: 'statbar-stat-label' });
-		charNoSpacesEl.createEl('span', { text: charNoSpaces.toLocaleString(), cls: 'statbar-stat-value' });
-
-		// Read time
-		const readTimeEl = statsContainer.createEl('div', { cls: 'statbar-stat-item' });
-		readTimeEl.createEl('span', { text: 'Estimated read time: ', cls: 'statbar-stat-label' });
-		readTimeEl.createEl('span', { text: `${readTime} minutes`, cls: 'statbar-stat-value' });
-
-		// Reading speed info
-		const speedEl = statsContainer.createEl('div', { cls: 'statbar-stat-item statbar-stat-info' });
-		speedEl.createEl('span', { text: `Based on ${this.plugin.settings.wordsPerMinute} words per minute`, cls: 'statbar-stat-note' });
-
-		// If there's a selection, also show document stats
-		if (isSelection) {
-			const fullWordCount = this.plugin.getWordCount(fullText);
-			const fullCharCount = fullText.length;
-			const fullReadTime = this.plugin.calculateReadTime(fullWordCount);
-
-			contentEl.createEl('hr');
-			contentEl.createEl('h3', { text: '📄 Full Document Statistics' });
-
-			const fullStatsContainer = contentEl.createEl('div', { cls: 'statbar-modal-stats' });
-
-			const fullWordEl = fullStatsContainer.createEl('div', { cls: 'statbar-stat-item' });
-			fullWordEl.createEl('span', { text: 'Words: ', cls: 'statbar-stat-label' });
-			fullWordEl.createEl('span', { text: fullWordCount.toLocaleString(), cls: 'statbar-stat-value' });
-
-			const fullCharEl = fullStatsContainer.createEl('div', { cls: 'statbar-stat-item' });
-			fullCharEl.createEl('span', { text: 'Characters: ', cls: 'statbar-stat-label' });
-			fullCharEl.createEl('span', { text: fullCharCount.toLocaleString(), cls: 'statbar-stat-value' });
-
-			const fullReadTimeEl = fullStatsContainer.createEl('div', { cls: 'statbar-stat-item' });
-			fullReadTimeEl.createEl('span', { text: 'Estimated read time: ', cls: 'statbar-stat-label' });
-			fullReadTimeEl.createEl('span', { text: `${fullReadTime} minutes`, cls: 'statbar-stat-value' });
-		}
-
-		// Close button
-		const buttonContainer = contentEl.createEl('div', { cls: 'statbar-modal-buttons' });
-		const closeButton = buttonContainer.createEl('button', { text: 'Close', cls: 'mod-cta' });
-		closeButton.addEventListener('click', () => this.close());
-	}
-
-	onClose() {
-		const { contentEl } = this;
-		contentEl.empty();
 	}
 }
