+++ src/types.ts
@@ -1,0 +1,7 @@
+export interface DocumentStats {
+	wordCount: number;
+	charCount: number;
+	readTime: string;
+	isSelection?: boolean;
+	selectionPrefix?: string;
+}
